Add Legion Context, Runtime args to Column, Table, etc constructors
The purpose is provide stable IndexSpaces and LogicalRegions associated with
Columns, Tables, and derived classes, which depend on Legion Context and Runtime
arguments to construct. Requiring those Legion value arguments to the
constructors allows Column, Table and derived class instances to contain values
of IndexSpace and LogicalRegion types. In most cases, IndexSpace and
LogicalRegion values are nevertheless created lazily.