Improve distinction of TableBuilder from reading a casacore table
The idea is for TableBuilders to be created as the user wants, from a casacore
table when desired. TableBuilders are generally useful, and creating one by
reading metadata from a casacore table is a special case. No longer support
providing a TableBuilder to a Table constructor directly, although methods on
TableBuilders may still be useful for providing some of the arguments of a Table
constructor.