Initialize fields in column physical region
Field values in the physical region for a table column that is being read are
uninitialized, which is problematic when assigning a value to those fields for
non-trivially copyable types. This is, apparently, primarily a problem for
casacore::String values. Although vector types are probably unaffected, I could
be wrong, and wouldn't rule out extending this solution to those types.