Change column indexing to eliminate all vector values
All casacore columns are now completely indexed by Legion IndexSpaces to scalar
values (subject to IndexSpace dimension limits). All index partitions of a
column can be projected to the index spaces of any other column in the same
table, eliminating the need for identical dimensions in all columns of a table.
This, however, leads to a requirement to enumerate the dimensions of every
column in the MS; it is left to a future revision to be able to access any
casacore table.