Allow empty Columns in a Table
This is necessary because casacore tables allow it. This means that in many
cases users must check that the return value of Column::index_space() is not
Legion::IndexSpace::NO_SPACE, and the return value of Column::logical_region()
is not Legion::LogicalRegion::NO_REGION.

TODO: allow user to prune empty columns from a Table?