Fix selection of call to casacore::Scalar/ArrayMeasColumn()
In get_meas_ref() we use a call to casacore::ScalarMeasColumn() or
casacore::ArrayMeasColumn() to get MeasRef instances; unfortunately, the naive
way to determine the selection of which of these two methods to call fails in
some cases (for example in t0.ms, table OBSERVATION, column TIME_RANGE). This
commit provides a workaround for that issue, but in a somewhat ungraceful way.
Enough dead code is retained to allow improvements at a later date.